+++ tests/Feature/InternetTypeDefinitions/EmailTest.php
@@ -1,9 +1,10 @@
 <?php
 
-
 use Redbox\Validation\Validator;
 
-dataset('invalid_emails', [
+dataset(
+    'invalid_emails',
+    [
     'test',
     'test@',
     'test@.com',
@@ -12,41 +13,59 @@ dataset('invalid_emails', [
     'test+test@.com',
     'test@%*.com',
     'test..',
-]);
+    ]
+);
 
-dataset('valid_emails', [
+dataset(
+    'valid_emails',
+    [
     'omething@something.com',
     'abc@gmail.com',
     'mastjohnny@gmail.com',
     'roadrunner@hotmail.com',
     'abc123@outlook.com',
-]);
+    ]
+);
 
-it('Validator::validate() should detect valid email addresses.', function ($email) {
+it(
+    'Validator::validate() should detect valid email addresses.',
+    function ($email) {
 
-    $validator = new Validator([
-        'email' => $email,
-    ]);
+        $validator = new Validator(
+            [
+            'email' => $email,
+            ]
+        );
 
-    $validator->validate([
-        'email' => 'email'
-    ]);
+        $validator->validate(
+            [
+            'email' => 'email'
+            ]
+        );
 
-    expect($validator->passes())->toBeTruthy()
+        expect($validator->passes())->toBeTruthy()
         ->and($validator->fails())->toBeFalsy();
-})->with('valid_emails');
+    }
+)->with('valid_emails');
 
 
-it('Validator::validate() should detect invalid email addresses.', function ($email) {
+it(
+    'Validator::validate() should detect invalid email addresses.',
+    function ($email) {
 
-    $validator = new Validator([
-        'email' => $email,
-    ]);
+        $validator = new Validator(
+            [
+            'email' => $email,
+            ]
+        );
 
-    $validator->validate([
-        'email' => 'email'
-    ]);
+        $validator->validate(
+            [
+            'email' => 'email'
+            ]
+        );
 
-    expect($validator->passes())->toBeFalsy()
+        expect($validator->passes())->toBeFalsy()
         ->and($validator->fails())->toBeTruthy();
-})->with('invalid_emails');
+    }
+)->with('invalid_emails');