+++ tests/Feature/InternetTypeDefinitions/IPV6Test.php
@@ -15,7 +15,9 @@ declare(strict_types=1);
 
 use Redbox\Validation\Validator;
 
-dataset('valid_ipv6_addresses', [
+dataset(
+    'valid_ipv6_addresses',
+    [
     '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
     '2001:db8:85a3:0:0:8a2e:370:7334',
     '2001:db8:85a3::8a2e:370:7334',
@@ -36,10 +38,13 @@ dataset('valid_ipv6_addresses', [
     '2001:db8:85a3:0:0:8A2E:370:7334',
     '2001:db8:85a3::8A2E:370:7334',
     '2001:db8:85a3:0:0:8A2E:370:7334',
-]);
+    ]
+);
 
 
-dataset('invalid_ipv6_addresses', [
+dataset(
+    'invalid_ipv6_addresses',
+    [
     '2001:0db8:85a3:0000:0000:8a2e:0370:7334:', // invalid
     '2001::85a3::8a2e::0370::7334', // invalid
     '2001:0db8:85a3:0000:0000:8a2e:0370:', // invalid
@@ -65,34 +70,49 @@ dataset('invalid_ipv6_addresses', [
     '2001:0db8:85a3:0000:0000:8a2e:0370:', // invalid
     '2001::85a3::8a2e::0370', // invalid
 
-]);
+    ]
+);
 
-it('Validator::validate() should detect valid ipv6 addresses.', function (string $ip) {
+it(
+    'Validator::validate() should detect valid ipv6 addresses.',
+    function (string $ip) {
 
-    $validator = new Validator([
-        'ip' => $ip,
-    ]);
+        $validator = new Validator(
+            [
+            'ip' => $ip,
+            ]
+        );
 
-    $validator->validate([
-        'ip' => 'ipv6'
+        $validator->validate(
+            [
+            'ip' => 'ipv6'
 
-    ]);
+            ]
+        );
 
-    expect($validator->passes())->toBeTruthy()
+        expect($validator->passes())->toBeTruthy()
         ->and($validator->fails())->toBeFalsy();
-})->with('valid_ipv6_addresses');
+    }
+)->with('valid_ipv6_addresses');
 
 
-it('Validator::validate() should detect invalid ipv6 addresses.', function (string $ip) {
+it(
+    'Validator::validate() should detect invalid ipv6 addresses.',
+    function (string $ip) {
 
-    $validator = new Validator([
-        'ip' => $ip,
-    ]);
+        $validator = new Validator(
+            [
+            'ip' => $ip,
+            ]
+        );
 
-    $validator->validate([
-        'ip' => 'ipv6'
-    ]);
+        $validator->validate(
+            [
+            'ip' => 'ipv6'
+            ]
+        );
 
-    expect($validator->passes())->toBeFalsy()
+        expect($validator->passes())->toBeFalsy()
         ->and($validator->fails())->toBeTruthy();
-})->with('invalid_ipv6_addresses');
+    }
+)->with('invalid_ipv6_addresses');