+++ tests/Feature/TypeDefinitions/BooleanTest.php
@@ -15,13 +15,19 @@ declare(strict_types=1);
 
 use Redbox\Validation\Validator;
 
-beforeEach(function () {
-    $validator = new Validator([
-        'field' => false,
-    ]);
-});
+beforeEach(
+    function () {
+        $validator = new Validator(
+            [
+            'field' => false,
+            ]
+        );
+    }
+);
 
-dataset('other_types_then_boolean', [
+dataset(
+    'other_types_then_boolean',
+    [
     null,
     "string",
     1,
@@ -31,20 +37,25 @@ dataset('other_types_then_boolean', [
     function () {
     },
     // Resource how ?
-]);
+    ]
+);
 
 
 test(
     'value false should be considered a boolean',
     function () {
 
-        $validator = new Validator([
-        'field' => false,
-        ]);
+        $validator = new Validator(
+            [
+            'field' => false,
+            ]
+        );
 
-        $validator->validate([
-        'field' => 'boolean'
-        ]);
+        $validator->validate(
+            [
+            'field' => 'boolean'
+            ]
+        );
 
         $errors = $validator->errors();
         expect(count($errors))->toEqual(0);
@@ -55,13 +66,17 @@ test(
     '[ALIAS] value false should be considered a bool',
     function () {
 
-        $validator = new Validator([
-        'field' => false,
-        ]);
+        $validator = new Validator(
+            [
+            'field' => false,
+            ]
+        );
 
-        $validator->validate([
-        'field' => 'bool'
-        ]);
+        $validator->validate(
+            [
+            'field' => 'bool'
+            ]
+        );
 
         $errors = $validator->errors();
         expect(count($errors))->toEqual(0);
@@ -72,13 +87,17 @@ test(
     'value true should be considered a boolean',
     function () {
 
-        $validator = new Validator([
-        'field' => false,
-        ]);
+        $validator = new Validator(
+            [
+            'field' => false,
+            ]
+        );
 
-        $validator->validate([
-        'field' => 'boolean'
-        ]);
+        $validator->validate(
+            [
+            'field' => 'boolean'
+            ]
+        );
 
         $errors = $validator->errors();
         expect(count($errors))->toEqual(0);
@@ -89,13 +108,17 @@ test(
     'running the boolean type check on a non-existing key in the target array should fail.',
     function () {
 
-        $validator = new Validator([
-        'field' => false,
-        ]);
+        $validator = new Validator(
+            [
+            'field' => false,
+            ]
+        );
 
-        $validator->validate([
-        'nonexisting' => 'boolean'
-        ]);
+        $validator->validate(
+            [
+            'nonexisting' => 'boolean'
+            ]
+        );
 
         expect($validator->fails())->toBeTruthy()
         ->and($validator->passes())->toBeFalsy();
@@ -106,20 +129,27 @@ test(
     }
 );
 
-test('Other types then integers should fail', function (mixed $type = null) {
-    $validator = new Validator([
-        'field' => $type,
-    ]);
+test(
+    'Other types then integers should fail',
+    function (mixed $type = null) {
+        $validator = new Validator(
+            [
+            'field' => $type,
+            ]
+        );
 
-    $validator->validate([
-        'field' => 'boolean'
-    ]);
+        $validator->validate(
+            [
+            'field' => 'boolean'
+            ]
+        );
 
-    expect($validator->passes())->toBeFalsy()
+        expect($validator->passes())->toBeFalsy()
         ->and($validator->fails())->toBeTruthy();
 
 
-    $errors = $validator->errors();
-    expect($errors['field'])->toEqual("Field field is not of type boolean.")
-        ->and(count($errors))->toEqual(1);
-})->with('other_types_then_boolean');
+        $errors = $validator->errors();
+        expect($errors['field'])->toEqual("Field field is not of type boolean.")
+            ->and(count($errors))->toEqual(1);
+    }
+)->with('other_types_then_boolean');