+++ tests/Feature/TypeDefinitions/IntergerTest.php
@@ -15,7 +15,9 @@ declare(strict_types=1);
 
 use Redbox\Validation\Validator;
 
-dataset('other_types_then_integer', [
+dataset(
+    'other_types_then_integer',
+    [
     null,
     true,
     "string",
@@ -25,19 +27,24 @@ dataset('other_types_then_integer', [
     function () {
     },
     // Resource how ?
-]);
+    ]
+);
 
 test(
     'value 10 should be considered a integer',
     function () {
 
-        $validator = new Validator([
-        'field' => 10,
-        ]);
+        $validator = new Validator(
+            [
+            'field' => 10,
+            ]
+        );
 
-        $validator->validate([
-        'field' => 'integer'
-        ]);
+        $validator->validate(
+            [
+            'field' => 'integer'
+            ]
+        );
 
         $errors = $validator->errors();
         expect(count($errors))->toEqual(0);
@@ -48,13 +55,17 @@ test(
     '[ALIAS] value 10 should be considered a int',
     function () {
 
-        $validator = new Validator([
-        'field' => 10,
-        ]);
+        $validator = new Validator(
+            [
+            'field' => 10,
+            ]
+        );
 
-        $validator->validate([
-        'field' => 'integer'
-        ]);
+        $validator->validate(
+            [
+            'field' => 'integer'
+            ]
+        );
 
         $errors = $validator->errors();
         expect(count($errors))->toEqual(0);
@@ -64,13 +75,17 @@ test(
 test(
     'Other types then integers should fail',
     function (mixed $type = null) {
-        $validator = new Validator([
-        'field' => $type,
-        ]);
+        $validator = new Validator(
+            [
+            'field' => $type,
+            ]
+        );
 
-        $validator->validate([
-        'field' => 'integer'
-        ]);
+        $validator->validate(
+            [
+            'field' => 'integer'
+            ]
+        );
 
         expect($validator->passes())->toBeFalsy()
         ->and($validator->fails())->toBeTruthy();
@@ -78,6 +93,6 @@ test(
 
         $errors = $validator->errors();
         expect($errors['field'])->toEqual("Field field is not of type integer.")
-        ->and(count($errors))->toEqual(1);
+            ->and(count($errors))->toEqual(1);
     }
 )->with('other_types_then_integer');