+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## v0.20
 
+* added type array 
 * added internet type ip
 * added internet type ipv4
 * added internet type ipv6
 * added internet type email
+* Fixed #10 where unknown types could be passed to validate and crash the validate function.
 
 ## v0.10
 
- * initial release
+* initial release